+++ packages/node/src/subcommands/reindex.module.ts
@@ -14,7 +14,10 @@ import {
   ConnectionPoolService,
   NodeConfig,
   ConnectionPoolStateManager,
+  IStoreModelProvider,
+  PlainStoreModelService,
 } from '@subql/node-core';
+import { Sequelize } from '@subql/x-sequelize';
 import { ConfigureModule } from '../configure/configure.module';
 import { ApiService } from '../indexer/api.service';
 import { DsProcessorService } from '../indexer/ds-processor.service';
@@ -23,7 +26,25 @@ import { UnfinalizedBlocksService } from '../indexer/unfinalizedBlocks.service';
 
 @Module({
   providers: [
-    StoreCacheService,
+    {
+      provide: 'IStoreModelProvider',
+      useFactory: (
+        nodeConfig: NodeConfig,
+        eventEmitter: EventEmitter2,
+        schedulerRegistry: SchedulerRegistry,
+        sequelize: Sequelize,
+      ): IStoreModelProvider => {
+        return nodeConfig.enableCache
+          ? new StoreCacheService(
+              sequelize,
+              nodeConfig,
+              eventEmitter,
+              schedulerRegistry,
+            )
+          : new PlainStoreModelService(sequelize, nodeConfig);
+      },
+      inject: [NodeConfig, EventEmitter2, SchedulerRegistry, Sequelize],
+    },
     StoreService,
     ReindexService,
     PoiService,